+++ packages/moonstone/ApplicationCloseDecorator/ApplicationCloseButton.js
@@ -8,20 +8,20 @@ import IconButton from '../IconButton';
 import css from './ApplicationCloseButton.less';
 
 /**
- * An {@link moonstone/ApplicationCloseButton.ApplicationCloseButton} with `closex` icon. It is used in
- * {@link moonstone/Panels.Panels} positioned at the top right corner.
+ * An {@link moonstone/ApplicationCloseDecorator.ApplicationCloseButton} with `closex` icon. It is
+ * positioned at the top right corner.
  * `onApplicationClose` callback function should be specified to close your app. The recommended
  * action to take with the event is `window.close()`, but you may also want to also do operations
  * like save user work or close database connections.
  *
  * @class ApplicationCloseButton
- * @memberof moonstone/Panels
+ * @memberof moonstone/ApplicationCloseDecorator
  * @private
  */
 const ApplicationCloseButton = kind({
 	name: 'ApplicationCloseButton',
 
-	propTypes: /** @lends moonstone/Panels.ApplicationCloseButton.prototype */ {
+	propTypes: /** @lends moonstone/ApplicationCloseDecorator.ApplicationCloseButton.prototype */ {
 		/**
 		 * A function to run when app close button is clicked
 		 *