+++ packages/moonstone/Panels/Panel.js
@@ -7,6 +7,8 @@ import Slottable from '@enact/ui/Slottable';
 import Spotlight from '@enact/spotlight';
 import SpotlightContainerDecorator, {spotlightDefaultClass} from '@enact/spotlight/SpotlightContainerDecorator';
 
+import {contextTypes} from '../ApplicationCloseDecorator';
+
 import css from './Panel.less';
 
 let panelId = 0;
@@ -44,6 +46,15 @@ const PanelBase = kind({
 		 * @public
 		 */
 		'aria-label': PropTypes.string,
+
+		/**
+		 * When `aria-owns` is set, it will add the given component(s) to the accessibility tree.
+		 *
+		 * @memberof moonstone/Panels.Panel.prototype
+		 * @type {String}
+		 * @public
+		 */
+		'aria-owns': PropTypes.string,
 
 		/**
 		 * Sets the strategy used to automatically focus an element within the panel upon render.
@@ -96,6 +107,8 @@ const PanelBase = kind({
 		noAutoFocus: PropTypes.bool
 	},
 
+	contextTypes,
+
 	defaultProps: {
 		autoFocus: 'last-focused',
 		hideChildren: false,
@@ -137,6 +150,19 @@ const PanelBase = kind({
 	},
 
 	computed: {
+		'aria-owns': ({'aria-owns': ariaOwns}, {hasCloseButton, closeButtonId}) => {
+			if (ariaOwns) {
+				// Close button added by ApplicationCloseDecorator is added to the panel's accessibility tree.
+				// This prevents re-reading the title when focus is moved from close button to the panel.
+				if (hasCloseButton) {
+					return `${ariaOwns} ${closeButtonId}`;
+				} else {
+					return ariaOwns;
+				}
+			} else if (hasCloseButton) {
+				return closeButtonId;
+			}
+		},
 		spotOnRender: ({autoFocus, hideChildren, noAutoFocus, spotOnRender}) => {
 			if (noAutoFocus) {
 				autoFocus = adaptToAutoFocus();