+++ packages/moonstone/Panels/Panels.js
@@ -1,12 +1,14 @@
+import deprecate from '@enact/core/internal/deprecate';
 import kind from '@enact/core/kind';
 import React from 'react';
 import PropTypes from 'prop-types';
 import {shape} from '@enact/ui/ViewManager';
 
+import ApplicationCloseButton from '../ApplicationCloseDecorator/ApplicationCloseButton';
+import {contextTypes} from '../ApplicationCloseDecorator';
 import IdProvider from '../internal/IdProvider';
 import Skinnable from '../Skinnable';
 
-import ApplicationCloseButton from './ApplicationCloseButton';
 import CancelDecorator from './CancelDecorator';
 import Viewport from './Viewport';
 
@@ -91,6 +93,7 @@ const PanelsBase = kind({
 		 * @type {Boolean}
 		 * @default false
 		 * @public
+		 * @deprecated replaced by `@enact/moonstone/ApplicationCloseDecorator`
 		 */
 		noCloseButton: PropTypes.bool,
 
@@ -99,6 +102,7 @@ const PanelsBase = kind({
 		 *
 		 * @type {Function}
 		 * @public
+		 * @deprecated replaced by `@enact/moonstone/ApplicationCloseDecorator/onApplicationClose`
 		 */
 		onApplicationClose: PropTypes.func,
 
@@ -110,6 +114,8 @@ const PanelsBase = kind({
 		 */
 		onBack: PropTypes.func
 	},
+
+	contextTypes,
 
 	defaultProps: {
 		index: 0,
@@ -126,9 +132,12 @@ const PanelsBase = kind({
 		className: ({noCloseButton, styler}) => styler.append({
 			hasCloseButton: !noCloseButton
 		}),
-		applicationCloseButton: ({id, noCloseButton, onApplicationClose}) => {
-			if (!noCloseButton) {
+		applicationCloseButton: ({id, noCloseButton, onApplicationClose}, {hasCloseButton}) => {
+			if (!hasCloseButton && !noCloseButton) {
 				const closeId = id ? `${id}_close` : null;
+
+				deprecate({name: 'noCloseButton', since: '1.5.0', replacedBy: '@enact/moonstone/ApplicationCloseDecorator'});
+				deprecate({name: 'onApplicationClose', since: '1.5.0', replacedBy: '@enact/moonstone/ApplicationCloseDecorator'});
 
 				return (
 					<ApplicationCloseButton
@@ -139,8 +148,10 @@ const PanelsBase = kind({
 				);
 			}
 		},
-		childProps: ({childProps, id, noCloseButton}) => {
-			if (noCloseButton || !id) {
+		childProps: ({childProps, id, noCloseButton}, {hasCloseButton}) => {
+			// If a close button is added by `ApplicationCloseDecorator`, `aria-owns` will be added in
+			// `moonstone/Panel` directly.
+			if (hasCloseButton || noCloseButton || !id) {
 				return childProps;
 			}