+++ packages/moonstone/Panels/tests/Panels-specs.js
@@ -44,6 +44,7 @@ describe('Panels Specs', () => {
 		expect(expected).to.equal(actual);
 	});
 
+	// TODO: remove following tests in 2.0.0
 	describe('computed', () => {
 		describe('childProps', () => {
 			it('should not add aria-owns when noCloseButton is true', () => {
@@ -54,9 +55,12 @@ describe('Panels Specs', () => {
 					noCloseButton: true,
 					id
 				};
+				const context = {
+					hasCloseButton: false
+				};
 
 				const expected = childProps;
-				const actual = PanelsBase.computed.childProps(props);
+				const actual = PanelsBase.computed.childProps(props, context);
 
 				expect(actual).to.equal(expected);
 			});
@@ -67,9 +71,12 @@ describe('Panels Specs', () => {
 					childProps,
 					noCloseButton: false
 				};
+				const context = {
+					hasCloseButton: false
+				};
 
 				const expected = childProps;
-				const actual = PanelsBase.computed.childProps(props);
+				const actual = PanelsBase.computed.childProps(props, context);
 
 				expect(actual).to.equal(expected);
 			});
@@ -82,9 +89,12 @@ describe('Panels Specs', () => {
 					noCloseButton: false,
 					id
 				};
+				const context = {
+					hasCloseButton: false
+				};
 
 				const expected = `${id}_close`;
-				const actual = PanelsBase.computed.childProps(props)['aria-owns'];
+				const actual = PanelsBase.computed.childProps(props, context)['aria-owns'];
 
 				expect(actual).to.equal(expected);
 			});
@@ -100,9 +110,50 @@ describe('Panels Specs', () => {
 					noCloseButton: false,
 					id
 				};
+				const context = {
+					hasCloseButton: false
+				};
 
 				const expected = `${ariaOwns} ${id}_close`;
-				const actual = PanelsBase.computed.childProps(props)['aria-owns'];
+				const actual = PanelsBase.computed.childProps(props, context)['aria-owns'];
+
+				expect(actual).to.equal(expected);
+			});
+		});
+
+		describe('when hasCloseButton is true', () => {
+			it('should not add aria-owns', () => {
+				const id = 'id';
+				const childProps = {};
+				const props = {
+					childProps,
+					noCloseButton: false,
+					id
+				};
+				const context = {
+					hasCloseButton: true
+				};
+
+				const expected = void 0;
+				const actual = PanelsBase.computed.childProps(props, context)['aria-owns'];
+
+				expect(actual).to.equal(expected);
+			});
+
+			it('should not render applicationCloseButton', () => {
+				const id = 'id';
+				const childProps = {};
+				const props = {
+					childProps,
+					noCloseButton: false,
+					id
+				};
+				const context = {
+					hasCloseButton: true
+				};
+
+				const expected = void 0;
+				const actual = PanelsBase.computed.childProps(props, context).applicationCloseButton;
 
 				expect(actual).to.equal(expected);
 			});